refactor: RandomNumberGameConstant 래팩토링
