--- src/main/java/sehee/domain/answer/AnswerConstant.java
@@ -1,7 +1,0 @@
-package sehee.domain.answer;
-
-public final class AnswerConstant {
-
-    public static final int RANDOM_NUMBER_LENGTH = 3;
-
-}
