+++ src/main/java/sehee/util/constant/RandomNumberGameConstant.java
@@ -1,0 +1,7 @@
+package sehee.util.constant;
+
+public final class RandomNumberGameConstant {
+
+    public static final int ANSWER_LENGTH = 3;
+
+}
